+++ tools/fixCoords/tool.js
@@ -1,14 +1,12 @@
 (()=>{
   class FixCoordsTool extends LiChessTools.Tools.ToolBase {
 
-    dependencies=['EmitRedraw'];
-
     preferences=[
       {
         name:'fixCoords',
         category: 'analysis',
         type:'multiple',
-        possibleValues: ['fix','larger'],
+        possibleValues: ['fix','larger','square'],
         defaultValue: 'fix',
         advanced: true
       }
@@ -19,15 +17,37 @@
         'options.analysis': 'Analysis',
         'options.fixCoords': 'Fix board coordinate position',
         'fixCoords.fix': 'Fix outside coordinates',
-        'fixCoords.larger': 'Larger coordinate font'
+        'fixCoords.larger': 'Larger coordinate font',
+        'fixCoords.square': 'On each square'
       },
       'ro-RO':{
         'options.analysis': 'Analiz\u0103',
         'options.fixCoords': 'Repar\u0103 pozi\u0163ia coordonatelor tablei',
         'fixCoords.fix': 'Repar\u0103 coordonatele in exterior',
-        'fixCoords.larger': 'Font mai mare pentru coordonate'
+        'fixCoords.larger': 'Font mai mare pentru coordonate',
+        'fixCoords.square': 'Pe fiecare p\u0103trat'
       }
     }
+
+    squareCoords=()=>{
+      const parent=this.lichessTools;
+      const lichess=parent.lichess;
+      const $=parent.$;
+      let container = $('div.main-board > div.cg-wrap > cg-container');
+      if (!container.length) return;
+      if (container.children('coords.lichessTools-fixCoords').length) return;
+      const coords=$('<coords class="lichessTools-fixCoords">')
+        .appendTo(container);
+      for (let rank=1; rank<=8; rank++) {
+        for (let file=1; file<=8; file++) {
+          $('<coord>')
+            .text(String.fromCharCode(96+rank)+String.fromCharCode(48+file))
+            .css('--rank',rank)
+            .css('--file',file)
+            .appendTo(coords);
+        }
+      }
+    };
 
     async start() {
       const parent=this.lichessTools;
@@ -37,12 +57,14 @@
       this.logOption('Fix coordinates', value);
       this.options={ 
         fix: parent.isOptionSet(value,'fix'),
-        larger: parent.isOptionSet(value,'larger')
+        larger: parent.isOptionSet(value,'larger'),
+        square: parent.isOptionSet(value,'square')
       };
       const body=$.cached('body');
       body
         .toggleClass('lichessTools-fixCoords-fix',this.options.fix)
-        .toggleClass('lichessTools-fixCoords-larger',this.options.larger);
+        .toggleClass('lichessTools-fixCoords-larger',this.options.larger)
+        .toggleClass('lichessTools-fixCoords-square',this.options.square);
       const analysis=lichess?.analysis;
       if (this._init_in===undefined) this._init_in=body.is('.coords-in');
       if (this._init_out===undefined) this._init_out=body.is('.coords-out');
@@ -57,6 +79,10 @@
         body
           .toggleClass('coords-in',this._init_in)
           .toggleClass('coords-out',this._init_out);
+      }
+      parent.global.clearInterval(this.interval);
+      if (this.options.square) {
+        this.interval=parent.global.setInterval(this.squareCoords,500);
       }
     }